requirements.txt
flask
flask-login
flask-sqlalchemy
pillow
pypdf2
selenium